- Use Int8 for intermediate oid count from alias file.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@26496 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/seqdb/seqdbalias.hpp
@@ -405,7 +405,7 @@ public:
     ///   The set of database volumes
     /// @return
     ///   The number of included sequences
-    int GetNumSeqs(const CSeqDBVolSet & volset) const;
+    Int8 GetNumSeqs(const CSeqDBVolSet & volset) const;
     
     /// Get the size of the OID range
     ///
@@ -418,7 +418,7 @@ public:
     ///   The set of database volumes
     /// @return
     ///   The number of OIDs found during traversal
-    int GetNumOIDs(const CSeqDBVolSet & volset) const;
+    Int8 GetNumOIDs(const CSeqDBVolSet & volset) const;
     
     /// Get the total length of the set of databases
     ///
@@ -877,7 +877,7 @@ public:
     ///   The set of database volumes
     /// @return
     ///   The number of included sequences
-    int GetNumSeqs(const CSeqDBVolSet & volset) const
+    Int8 GetNumSeqs(const CSeqDBVolSet & volset) const
     {
         return m_Node.GetNumSeqs(volset);
     }
@@ -893,7 +893,7 @@ public:
     ///   The set of database volumes
     /// @return
     ///   The number of OIDs found during traversal
-    int GetNumOIDs(const CSeqDBVolSet & volset) const
+    Int8 GetNumOIDs(const CSeqDBVolSet & volset) const
     {
         return m_Node.GetNumOIDs(volset);
     }
